Added More Mongo Fucntions
The start of the Employee Login is no implemented. Error codes are not returned to the end user. Now using Interface for user sessions in order to allow for the App to check for both Employee and Manager Sessions in a single query.

Still need to work on error values being returned. Login does not yet create a session